Fixed Migrations. Increase users model. Add new styles and scripts. Implemented "My profile". Add posibility to change password

+++ TestOk/BusinessLogic/Services/SurveyService.cs
@@ -58,7 +58,7 @@ namespace BusinessLogic.Services
 
             var quizes = survey.Test.Quizes.Select(q => _quizRepository.GetQuizById(q.Id));
 
-            decimal mark = 0;
+            double mark = 0;
 
             foreach (var quiz in quizes)
             {
