+++ TestOk/DataAccess/Data/Models/Quiz.cs
@@ -6,8 +6,8 @@ namespace DataAccess.Data.Models
     {
         public int Id { get; set; }
         public string Question { get; set; }
-        public decimal PointsPerCorrectAnswer { get; set; }
-        public decimal Complexity { get; set; }
+        public double PointsPerCorrectAnswer { get; set; }
+        public double Complexity { get; set; }
         public List<QuizOption> Options { get; set; }
         public List<QuizOption> CorrectAnswers { get; set; }
     }
